Move: Automlstub thing to it's own fixture
