+++ test/conftest.py
@@ -1,117 +1,23 @@
-from typing import Any, Iterator, List, Optional, Callable
+"""
+Testing
+=======
 
-import os
+**Features**
+* marker - ``pytest.mark.todo``` to mark a test which xfails as it's todo
+* fixtures - All fixtures in "test/fixtures" are known in every test file
+"""
+from typing import Iterator, List, Optional
+
 import re
-import shutil
-import time
-import unittest.mock
 from pathlib import Path
 
 import psutil
 import pytest
-from dask.distributed import Client, get_client
 from pytest import ExitCode, FixtureRequest, Item, Session
 
-from autosklearn.automl import AutoML
-from autosklearn.automl_common.common.utils.backend import Backend, create
 
 HERE = Path(__file__)
 
-
-class AutoMLStub(AutoML):
-    def __init__(self) -> None:
-        self.__class__ = AutoML
-        self._task = None
-        self._dask_client = None
-        self._is_dask_client_internally_created = False
-
-    def __del__(self) -> None:
-        pass
-
-
-@pytest.fixture(scope="function")
-def automl_stub(request: FixtureRequest) -> AutoMLStub:
-    automl = AutoMLStub()
-    automl._seed = 42
-    automl._backend = unittest.mock.Mock(spec=Backend)
-    automl._backend.context = unittest.mock.Mock()
-    automl._delete_output_directories = lambda: 0
-    return automl
-
-
-@pytest.fixture(scope="function")
-def backend(request: FixtureRequest) -> Backend:
-    test_dir = os.path.dirname(__file__)
-    tmp = os.path.join(
-        test_dir, ".tmp__%s__%s" % (request.module.__name__, request.node.name)
-    )
-
-    for dir in (tmp,):
-        for i in range(10):
-            if os.path.exists(dir):
-                try:
-                    shutil.rmtree(dir)
-                    break
-                except OSError:
-                    time.sleep(1)
-
-    # Make sure the folders we wanna create do not already exist.
-    backend = create(
-        temporary_directory=tmp, output_directory=None, prefix="auto-sklearn"
-    )
-
-    def get_finalizer(tmp_dir: str) -> Callable:
-        def session_run_at_end() -> None:
-            for dir in (tmp_dir,):
-                for i in range(10):
-                    if os.path.exists(dir):
-                        try:
-                            shutil.rmtree(dir)
-                            break
-                        except OSError:
-                            time.sleep(1)
-
-        return session_run_at_end
-
-    request.addfinalizer(get_finalizer(tmp))
-
-    return backend
-
-
-@pytest.fixture(scope="function")
-def tmp_dir(request: FixtureRequest) -> str:
-    return _dir_fixture("tmp", request)
-
-
-def _dir_fixture(dir_type: str, request: FixtureRequest) -> str:
-    test_dir = os.path.dirname(__file__)
-
-    dirname = f".{dir_type}__{request.module.__name__}__{request.node.name}"
-    dir = os.path.join(test_dir, dirname)
-
-    for i in range(10):
-        if os.path.exists(dir):
-            try:
-                shutil.rmtree(dir)
-                break
-            except OSError:
-                pass
-
-    def get_finalizer(dir: str) -> Callable:
-        def session_run_at_end() -> None:
-            for i in range(10):
-                if os.path.exists(dir):
-                    try:
-                        shutil.rmtree(dir)
-                        break
-                    except OSError:
-                        time.sleep(1)
-
-        return session_run_at_end
-
-    request.addfinalizer(get_finalizer(dir))
-
-    return dir
 
 
 def walk(path: Path, include: Optional[str] = None) -> Iterator[Path]:
@@ -158,9 +64,6 @@ def fixture_modules() -> List[str]:
     ]
 
 
-pytest_plugins = fixture_modules()
-
-
 def pytest_runtest_setup(item: Item) -> None:
     """Run before each test"""
     todos = [mark for mark in item.iter_markers(name="todo")]
@@ -172,3 +75,6 @@ def pytest_sessionfinish(session: Session, exitstatus: ExitCode) -> None:
     proc = psutil.Process()
     for child in proc.children(recursive=True):
         print(child, child.cmdline())
+
+
+pytest_plugins = fixture_modules()
